requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
pandas==2.2.0
selenium==4.25.0
streamlit==1.41.0
google-generativeai==0.8.0
chromadb==0.5.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
nest-asyncio==1.6.0
